+++ app/helpers/songs_helper.rb
@@ -40,20 +40,6 @@ module SongsHelper
         end
     end
 
-    # returns the correct filter from a given song_filter_id (HTML id)
-    def song_filter_from_id(song_filter_id, current_sort = :none, user_id: nil)
-        case song_filter_id
-        when :sort_by_capo.to_s.kebabcase
-            song_filter_capo(current_sort)
-        when :sort_by_most_played_by_user.to_s.kebabcase
-            song_filter_most_played_by_user(user_id, current_sort)
-        when :sort_by_last_played.to_s.kebabcase
-            song_filter_last_played(user_id, current_sort)
-        when :sort_by_created_at.to_s.kebabcase
-            song_filter_created_at(current_sort)
-        end
-    end
-
     # a button that lets user filter songs by their suggested ca
     # @current_sort decides how the button looks, to indicate the user the current sort, default to :desc so that it initializes with no sort, as :desc sort leads to no sort (see rotate_filter_options)
     def song_filter_capo(current_sort = :none)